+++ R/adsl.R
@@ -3,7 +3,7 @@
 #' @description adsl modified from pharmaverseadam
 #' @source data from pharmaverseadam.
 #'
-#' @format A data frame with 306 rows and 116 variables:
+#' @format A data frame with 306 rows and 119 variables:
 #' \describe{
 #'  \item{STUDYID}{Study Identifier}
 #'  \item{USUBJID}{Unique Subject Identifier}
@@ -121,10 +121,13 @@
 #'  \item{PPREXRS}{Reason for Excl from Per-Prot Population}
 #'  \item{PKEXRES}{Reason for Excl from Pharmacokin Pop}
 #'  \item{IMEXRES}{Reason for Excl from Immunogen Pop}
+#'  \item{COHORT}{Cohort}
+#'  \item{GROUP}{Analysis Group}
+#'  \item{EOTDT}{End-of-Treatment Date}
 #' }
 #' @seealso \code{\link{adae}} \code{\link{adaeocmq}} \code{\link{adagocmq}} \code{\link{adcm}} \code{\link{addili}} \code{\link{adeg}} \code{\link{adex}} \code{\link{adexsum}} \code{\link{adlb}} \code{\link{adpc}} \code{\link{adsl}} \code{\link{adttesaf}} \code{\link{advs}}# nolint
 #' @keywords datasets adsl
 #' @name adsl
 #' @examples
-#' head(data("adsl"))
+#'  head(data("adsl"))
 "adsl"