+++ data-raw/adsl.R
@@ -407,6 +407,49 @@ gen_adsl <- function(seed = 123) {
     ),
   )
 
+  gen <- dplyr::mutate(
+    gen,
+    DCSCREEN = case_when(
+      USUBJID == "01-701-1240" ~ "Subject refused to sign informed consent",
+      .default = DCSCREEN
+    ),
+    RESCRNFL = if_else(
+      SCRFFL == "Y" & runif(n()) < 0.5, "Y", NA_character_
+    )
+  )
+
+
+  gen <- gen |>
+    dplyr::mutate(
+      COHORT = factor(
+        dplyr::case_match(
+          ARM,
+          "Placebo" ~ "Cohort 1",
+          "Xanomeline High Dose" ~ "Cohort 2",
+          "Xanomeline Low Dose" ~ "Cohort 3",
+          "Screen Failure" ~ NA_character_,
+          .default = NA_character_
+        ),
+
+        levels = c("Cohort 1", "Cohort 2", "Cohort 3")
+      ),
+
+      GROUP = factor(
+        dplyr::case_match(
+          ARM,
+          "Placebo" ~ "Group 1",
+          "Xanomeline High Dose" ~ "Group 2",
+          "Xanomeline Low Dose" ~ "Group 3",
+          "Screen Failure" ~ NA_character_,
+          .default = NA_character_
+        ),
+
+        levels = c("Group 1", "Group 2", "Group 3")
+      ),
+
+      EOTDT = TRTEDT
+    )
+
   # Define additional labels for new variables not in source dataset
   additional_labels <- list(
     TRT01PN = "Planned Treatment for Period 01 (N)",
@@ -472,7 +515,11 @@ gen_adsl <- function(seed = 123) {
     FASEXRS = "Reason for Excl from Full Analysis Set",
     PPREXRS = "Reason for Excl from Per-Prot Population",
     PKEXRES = "Reason for Excl from Pharmacokin Pop",
-    IMEXRES = "Reason for Excl from Immunogen Pop"
+    IMEXRES = "Reason for Excl from Immunogen Pop",
+    COHORT = "Cohort",
+    GROUP = "Analysis Group",
+    EOTDT = "End-of-Treatment Date",
+    BRTHDTC = "Date/Time of Birth"
   )
 
   # Handle NA values and convert characters to factors