+++ data-raw/create_data.R
@@ -21,7 +21,7 @@ data_scripts <- list.files(
   pattern = "\\.R$",
   full.names = TRUE
 )
-
+data_scripts<- data_scripts[1]
 # Filter out helper and create_data scripts
 data_scripts <- data_scripts[
   !grepl("(helpers\\.R|create_data\\.R)", data_scripts)
@@ -151,5 +151,5 @@ walk(data_rda, run_xpt)
 
 message("All datasets have been transformed.")
 
-system("air format .")
+system("air format data-raw/")
 message("All datasets have been formated.")